Reduced number of tiles, created a test map for reading, switched map from int[][] to short[][] to save memory. This will be the case until we discover that we need more than 65535 tiles.

+++ src/PokemonGen3/Map.java
@@ -35,13 +35,13 @@ public class Map {
     private IntegerBinding playerPositionY;
 
     
-    private int[][] map;
+    private short[][] map;
     
     /**
      * Creates a map. Can be easily manipulated using a MapHelper
      * @param startingMap 
      */
-    public Map(int[][] startingMap)
+    public Map(short[][] startingMap)
     {
         map = startingMap;
         height = map.length;
